requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
--index-url https://pypi.tuna.tsinghua.edu.cn/simple

aiomysql==0.0.20          # via databases
aioredis==1.3.1
apispec-webframeworks==0.5.2
apispec[yaml]==3.2.0      # via apispec-webframeworks
async-timeout==3.0.1      # via aioredis
cffi==1.13.2              # via cryptography
cryptography==2.8         # via pymysql
databases[mysql]==0.2.6
hiredis==1.0.1            # via aioredis
marshmallow==3.3.0
pycparser==2.19           # via cffi
pymysql==0.9.2            # via aiomysql, databases
python-dateutil==2.8.1
pyyaml==5.2               # via apispec
raven==6.10.0
six==1.13.0               # via cryptography, python-dateutil
sqlalchemy==1.3.12        # via databases
tornado==6.0.3
uvloop==0.14.0
